Moved postings published from wrong wave to shared wave

+++ lib/tasks/refresh.rake
@@ -28,32 +28,35 @@ namespace :ff do
     task :load => [ :'spec:db:fixtures:load', :'spec:db:fixtures:images:load', :'^^db:seed' ] # :'ts:rebuild'
   end
 
-  namespace :db do    
-    desc "Refresh development with production db and images from local dumps. Use DUMP_DATE=yyyymmdd"
-    task :refresh => [ :'refresh:db', :'refresh:images', :'refresh:default_wave' ]
+  namespace :db do
+    desc "Set the default wave"
+    task :default_wave => :environment do      
+      wave = Wave::Base.find_by_slug('shared')
+      if wave.present?
+        wave.slug = ::WavesController::DefaultWaveSlug
+        wave.save
+      end
+    end
+    
+    desc "Refresh development with production db and images from local dumps. DUMP_DATE=yyyymmdd"
+    task :refresh => [ :'refresh:sql', :'refresh:images', :default_wave ]
   
     namespace :refresh do
-      task :db => :environment do
+      desc "Refresh development with production sql from local. DUMP_DATE=yyyymmdd"
+      task :sql => :environment do
         require 'yaml'
         database  = YAML::load_file("config/database.yml")
         dump_filename = "dump.#{timestamp}.sql"
         sh "mysql -u #{database['development']['username']} -p#{database['development']['password']} #{database['development']['database']} < db/dumps/#{dump_filename}"
       end
     
-      desc "Refresh development with production images from local tar dump. Use DUMP_DATE=yyyymmdd"
+      desc "Refresh development with production images from local tar dump. DUMP_DATE=yyyymmdd"
       task :images => :environment do
         tar_filename = "images.#{timestamp}.tar.gz"
-        sh "rm -rf public/system/images/* && tar -xf db/dumps/#{tar_filename} -C public/system"
-      end
-      
-      desc "Set the default wave"
-      task :default_wave => :environment do      
-        wave = Wave::Base.find_by_slug('shared')
-        if wave.present?
-          wave.slug = ::WavesController::DefaultWaveSlug
-          wave.save
+        if File.exists?("db/dumps/#{tar_filename}")
+          sh "rm -rf public/system/images/* && tar -xf db/dumps/#{tar_filename} -C public/system"
         end
-      end
+      end      
     end
   end
 end
